requirements.txt
openai
flask
python-dotenv
mysql-connector-python
flask_login
gunicorn
google-generativeai


